window: Use guint8 for opacity internally
Except while reading _NET_WM_WINDOW_OPACITY, opacity is between 0 and 255. With
guint8, we'll get compiler warnings if arbitrary int values are passed.

https://bugzilla.gnome.org/show_bug.cgi?id=727874

+++ src/core/window-private.h
@@ -370,8 +370,8 @@ struct _MetaWindow
   /* the input shape region for picking */
   cairo_region_t *input_region;
 
-  /* _NET_WM_WINDOW_OPACITY */
-  guint opacity;
+  /* _NET_WM_WINDOW_OPACITY rescaled to 0xFF */
+  guint8 opacity;
 
   /* if TRUE, the we have the new form of sync request counter which
    * also handles application frames */
@@ -731,7 +731,7 @@ void meta_window_set_transient_for        (MetaWindow *window,
                                            MetaWindow *parent);
 
 void meta_window_set_opacity              (MetaWindow *window,
-                                           guint       opacity);
+                                           guint8      opacity);
 
 void meta_window_set_custom_frame_extents (MetaWindow *window,
                                            GtkBorder  *extents);
